Organize notebook related services
