+++ src/vs/workbench/contrib/notebook/browser/services/notebookCellStatusBarServiceImpl.ts
